refactor(product): extract useHomeComposerAttachments; refresh sidebar structure entry (PRO-165)
HomeNextScreen crossed the ~400-line frontend split threshold once the
drop wiring landed; the attachment controller + drag/drop handlers move to
a home ui hook. MainSidebar.tsx's structure-allowlist entry refreshes to
the observed 524 after #1856's main-side growth.
