+++ apps/packages/product-client/src/components/home/screen/HomeNextScreen.tsx
@@ -1,5 +1,4 @@
-import { useCallback, useMemo, useRef, useState, type DragEvent } from "react";
-import { useProductHost } from "@proliferate/product-client/host/ProductHostProvider";
+import { useRef, useState } from "react";
 import { HomeComposerForm } from "#product/components/home/screen/HomeComposerForm";
 import { HomeOnboardingCards } from "#product/components/home/screen/HomeOnboardingCards";
 import { HomeProjectMenu } from "#product/components/home/screen/HomeProjectMenu";
@@ -15,20 +14,15 @@ import {
 import { DebugProfiler } from "#product/components/diagnostics/DebugProfiler";
 import { Button } from "#product/primitives/Button";
 import { Input } from "#product/primitives/Input";
-import { useChatPromptAttachments } from "#product/hooks/chat/ui/use-chat-prompt-attachments";
+import { useHomeComposerAttachments } from "#product/hooks/home/ui/use-home-composer-attachments";
 import { useHomeNextLaunchControls } from "#product/hooks/home/derived/use-home-next-launch-controls";
 import { useHomeCloudRepoSettingsNavigation } from "#product/hooks/home/workflows/use-home-cloud-repo-settings-navigation";
 import { useHomeNextTargetSelectionState } from "#product/hooks/home/ui/use-home-next-target-selection-state";
 import { useHomeNextState } from "#product/hooks/home/derived/use-home-next-state";
 import { useHomeScreen } from "#product/hooks/home/facade/use-home-screen";
 import {
-  isFileDrag,
-  readFileDragInput,
-} from "#product/lib/domain/chat/composer/prompt-attachment-drag";
-import {
   buildHomeModelSelectorProps,
   buildHomeSessionConfigControls,
-  HOME_COMPOSER_PROMPT_CAPABILITIES,
 } from "#product/lib/domain/home/home-composer-controls";
 import { type HomeNextModelSelection } from "#product/lib/domain/home/home-next-launch";
 import { resolveHomeModelProbeCardState } from "#product/lib/domain/home/home-screen";
@@ -86,89 +80,14 @@ export function HomeNextScreen() {
   // home-scoped controller with optimistic pre-session capabilities and ride
   // the launch as prompt snapshots — see useHomeNextComposerState.submit).
   const homeAgentKind = homeNext.effectiveModelSelection?.kind ?? null;
-  const host = useProductHost();
-  const desktopFiles = host.desktop?.files ?? null;
-  // Dropped-path recovery mirrors ChatView: local path references are only
-  // meaningful when the launched agent will share this machine's filesystem.
-  // cowork/local/worktree run here; cloud does not, and an unresolved target
-  // (including ssh) gets no local refs either. The attachment scope keys off
-  // the same split so flipping the target across the local/remote boundary
-  // clears drafts instead of letting a stale local path ride into a runtime
-  // that cannot read it.
-  const launchTargetKind = homeNext.launchTarget?.kind ?? null;
-  const isLocalRuntimeTarget = launchTargetKind === "cowork"
-    || launchTargetKind === "local"
-    || launchTargetKind === "worktree";
-  // One promise per drag session (see ChatView): a promise's value cannot be
-  // clobbered by a later session's capture, and a drop that lands before the
-  // capture resolves awaits it instead of skipping the comparison.
-  const dragSessionChangeCountRef = useRef<Promise<number> | null>(null);
-  const pendingDropChangeCountRef = useRef<Promise<number> | null>(null);
-  const resolveDroppedPaths = useMemo(() => {
-    if (!desktopFiles || !isLocalRuntimeTarget) {
-      return null;
-    }
-    return async () => {
-      const expectedChangeCount = pendingDropChangeCountRef.current;
-      pendingDropChangeCountRef.current = null;
-      // No captured drag session means the snapshot cannot be attributed;
-      // empty entries route the drop to byte uploads.
-      if (!expectedChangeCount) {
-        return [];
-      }
-      const [snapshot, expected] = await Promise.all([
-        desktopFiles.readDroppedPaths(),
-        expectedChangeCount,
-      ]);
-      // A different drag session wrote the pasteboard after this drop's drag
-      // entered the surface.
-      return snapshot.changeCount === expected ? snapshot.entries : [];
-    };
-  }, [desktopFiles, isLocalRuntimeTarget]);
-  const attachments = useChatPromptAttachments({
-    scopeKey: isLocalRuntimeTarget ? "home:local" : "home:remote",
-    promptCapabilities: HOME_COMPOSER_PROMPT_CAPABILITIES,
-    canAttachFiles: true,
-    resolveDroppedPaths,
-  });
+  const {
+    attachments,
+    fileDragOver,
+    handleFileDrag,
+    handleDrop,
+    handleDragLeave,
+  } = useHomeComposerAttachments(homeNext.launchTarget?.kind ?? null);
   const fileInputRef = useRef<HTMLInputElement>(null);
-  const [fileDragOver, setFileDragOver] = useState(false);
-  const addDroppedFiles = attachments.addDroppedFiles;
-  const handleFileDrag = useCallback((event: DragEvent<HTMLDivElement>) => {
-    if (!isFileDrag(readFileDragInput(event.dataTransfer))) {
-      return;
-    }
-    event.preventDefault();
-    event.dataTransfer.dropEffect = "copy";
-    setFileDragOver(true);
-    if (dragSessionChangeCountRef.current === null && desktopFiles && resolveDroppedPaths) {
-      // Arm once per drag session; the count identifies the session that
-      // will deliver the drop.
-      dragSessionChangeCountRef.current = desktopFiles.getDragPasteboardChangeCount();
-    }
-  }, [desktopFiles, resolveDroppedPaths]);
-  const handleDrop = useCallback((event: DragEvent<HTMLDivElement>) => {
-    if (!isFileDrag(readFileDragInput(event.dataTransfer))) {
-      return;
-    }
-    event.preventDefault();
-    event.stopPropagation();
-    setFileDragOver(false);
-    const sessionChangeCount = dragSessionChangeCountRef.current;
-    dragSessionChangeCountRef.current = null;
-    pendingDropChangeCountRef.current = sessionChangeCount;
-    // No files.length gate: WebKit can surface folder-only drops with an
-    // empty FileList, and the host path resolver still recovers those items.
-    addDroppedFiles(event.dataTransfer.files);
-  }, [addDroppedFiles]);
-  const handleDragLeave = useCallback((event: DragEvent<HTMLDivElement>) => {
-    const relatedTarget = event.relatedTarget;
-    if (relatedTarget instanceof Node && event.currentTarget.contains(relatedTarget)) {
-      return;
-    }
-    setFileDragOver(false);
-    dragSessionChangeCountRef.current = null;
-  }, []);
   const homeSessionConfigControls = buildHomeSessionConfigControls({
     destination,
     agentKind: homeAgentKind,
